Add release suites to stdeb.cfg.
When releasing with packagecloud the Suites field of the Debian changes
file was not used for controlling release suites and was removed. Now
that we are again releasing directly into the ROS bootstrap repository
our tools do use this field again.

Start by releasing into actively supported Ubuntu and Debian
distributions and expand support to older distributions only if needed.