Configuration updates for stdeb release to bootstrap repository. (#25)
* Conflict our new vcs2l package with existing python3-vcstool installations

This is how we handle the python2 / python3 packages that both install
the same tool to /usr/bin so it seems reasonable to do for vcstool vs
vcs2l.

* Add release suites to stdeb.cfg.

When releasing with packagecloud the Suites field of the Debian changes
file was not used for controlling release suites and was removed. Now
that we are again releasing directly into the ROS bootstrap repository
our tools do use this field again.

Start by releasing into actively supported Ubuntu and Debian
distributions and expand support to older distributions only if needed.